- second layer for test created. Layer contains ApplicationManager that delegates its funtionality to Helpers - test are refactored due to new test suite architecture
