+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/appmanager/NavigationHelper.java
@@ -1,0 +1,20 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+public class NavigationHelper {
+  private FirefoxDriver wd;
+
+  public NavigationHelper(FirefoxDriver wd) {
+    this.wd = wd;
+  }
+
+  public void goToGroupPage() {
+    wd.findElement(By.linkText("groups")).click();
+  }
+
+  public void goToHomePage() {
+    wd.findElement(By.linkText("home")).click();
+  }
+}
