+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/appmanager/SessionHelper.java
@@ -1,0 +1,22 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+public class SessionHelper {
+  private FirefoxDriver wd;
+
+  public SessionHelper(FirefoxDriver wd) {
+    this.wd = wd;
+  }
+
+  public void login(String username, String password) {
+    wd.findElement(By.name("user")).click();
+    wd.findElement(By.name("user")).clear();
+    wd.findElement(By.name("user")).sendKeys(username);
+    wd.findElement(By.name("pass")).click();
+    wd.findElement(By.name("pass")).clear();
+    wd.findElement(By.name("pass")).sendKeys(password);
+    wd.findElement(By.xpath("//form[@id='LoginForm']/input[3]")).click();
+  }
+}
