+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/tests/ContactCreationTests.java
@@ -8,10 +8,10 @@ public class ContactCreationTests extends TestBase {
   @Test
   public void testContactCreation() {
 
-    app.initContactCreation();
-    app.fillContactForm(new ContactData("Serhiy", "Tymoshenko", "17. Sichovy str.\n116 apt", "serdtimosh@gmail.com", "+0472313950", "+380945643839"));
-    app.submitContactCreation();
-    app.goToHomePage();
+    app.getContactHelper().initContactCreation();
+    app.getContactHelper().fillContactForm(new ContactData("Serhiy", "Tymoshenko", "17. Sichovy str.\n116 apt", "serdtimosh@gmail.com", "+0472313950", "+380945643839"));
+    app.getContactHelper().submitContactCreation();
+    app.getNavigationHelper().goToHomePage();
 
   }
